Fix test - container is created by no longer crashes on parse error because of named args
