+++ tests/PHPStan/Analyser/DynamicMethodThrowTypeExtensionTest.php
@@ -15,7 +15,7 @@ class DynamicMethodThrowTypeExtensionTest extends TypeInferenceTestCase
 		}
 
 		yield from $this->gatherAssertTypes(__DIR__ . '/data/dynamic-method-throw-type-extension.php');
-		yield from $this->gatherAssertTypes(__DIR__ . '/data/dynamic-method-throw-type-extension-named-args.php');
+		yield from $this->gatherAssertTypes(__DIR__ . '/data/dynamic-method-throw-type-extension-named-args-fixture.php');
 	}
 
 	/**
